+++ renderlib/AppearanceObject.cpp
