+++ renderlib/AppearanceObject.hpp
