+++ renderlib/CMakeLists.txt
@@ -16,8 +16,8 @@ target_include_directories(renderlib PUBLIC
 target_sources(renderlib PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/AppearanceDataObject.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/AppearanceDataObject.hpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/AppearanceUiDescription.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/AppearanceUiDescription.hpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/AppearanceObject.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/AppearanceObject.hpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/AppScene.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/AppScene.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/AreaLightTool.cpp"
@@ -30,8 +30,8 @@ target_sources(renderlib PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/CCamera.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/CameraDataObject.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/CameraDataObject.hpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/CameraUiDescription.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/CameraUiDescription.hpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/CameraObject.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/CameraObject.hpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/ClipPlaneTool.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/ClipPlaneTool.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Colormap.cpp"
