+++ renderlib/CameraObject.cpp
