+++ renderlib/CameraObject.hpp
